update: replace the action with api route handler
